+++ CHANGELOG.md
@@ -4,9 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Modern Some Web Clicker [0.6.0 - 0.9.0]
-### Unreleased [10/25/2018]
+## Modern Some Web Clicker [0.6.0 - 0.10.0]
+### 0.10.0 [10/26/2018]
 #### Additions
+- **NEW!** Click Miner upgrade.
 - Added click and buy sounds that only play on desktops. This is because there is a delay on mobile that lags the game.
 
 #### Known Bugs